apps/telegram/requirements.txt
fastapi
uvicorn
telethon
pydantic
python-dotenv
pillow
